% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeroCasosPorGrupoEstagiamento.R
\name{numeroCasosPorGrupoEstagiamento}
\alias{numeroCasosPorGrupoEstagiamento}
\title{Title}
\usage{
numeroCasosPorGrupoEstagiamento(...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
