% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeroCasosPorLocalizacaoAgrupada.R
\name{numeroCasosPorLocalizacaoAgrupada}
\alias{numeroCasosPorLocalizacaoAgrupada}
\title{Title}
\usage{
numeroCasosPorLocalizacaoAgrupada(...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
