% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeroCasosPorLocalizacaoPrimaria.R
\name{numeroCasosPorLocalizacaoPrimaria}
\alias{numeroCasosPorLocalizacaoPrimaria}
\title{Title}
\usage{
numeroCasosPorLocalizacaoPrimaria(...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
